@@ -1,6 +1,6 @@
 {
   "name": "@nexushubs/react-simple-di",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "Simple Dependancy Injection Solution for React",
   "repository": {
     "type": "git",
@@ -38,10 +38,12 @@
     "enzyme": "^2.2.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0 || ^16.0.0"
+    "react": "^16.3.0",
+    "create-react-class": "^15.6.3",
+    "prop-types": "^15.6.1"
   },
   "dependencies": {
     "babel-runtime": "6.x.x",
-    "hoist-non-react-statics": "1.x.x"
+    "hoist-non-react-statics": "2.5.x"
   }
 }
